refactor conkit.command_line.cli_tools
